requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
SQLAlchemy==2.0.45
python-dotenv
psycopg2-binary==2.9.11
pydantic
alembic
passlib
bcrypt
pyjwt==2.10.0